ocr_microservice/requirements.txt
fastapi
uvicorn[standard]
python-multipart
transformers[torch]
torch
accelerate
Pillow
pytesseract
python-dotenv